bram integration
First try of bram integration is finished. Delay data can be stored in the bram, and verified to be fairly consistent. Certain changes are made to PUF_Function.vhd, to make the delay stored in the bram be more consistent. For future design, maybe need to add an extra file for bram control only.